refactor(interpreter): extract redirection handling into submodule
Move process_input_redirections and the apply_redirections / fd-routing core
(apply_redirections, apply_redirections_fd_table, logic_only_redirect_error)
out of interpreter/mod.rs into interpreter/redirection.rs (~440 lines). The
FdTarget enum and route_fd_table_content helper stay in the parent (referenced
by interpreter state fields).
